bytes, strings: fix bad worst-case behavior in Index and use bytealg.Index to look for prefixes of long values of sep
The worst-case behavior was described in a TODO in the old code: a long
prefix of sep matches frequently when trying IndexByte, causing runtime to
grow with len(sep)*len(s). The change here is to check whether a mismatch
happened >64 bytes into sep and, if so, add a 'penalty' of len(sep)/64 to
the failure counter. That breaks out of the loop on worst-case data, but
is meant to be conservative enough it will rarely change anything for
well-behaved inputs. There's a benchmark showing the slowdown with the old
code and that the change fixes it.

When n > bytealg.MaxLen and IndexByte fails, we'll now try bytealg.Index on
a prefix of sep, then check if there's a full match, before falling back to
indexRabinKarp. We bail out if two false prefix matches are less than
len(sep) apart. It takes a specific input pattern to trigger this, but when
it does trigger it can help a lot, 68% on Skylake amd64. There's a benchmark
demonstrating that, and a couple inputs designed to be bad for this change
to show it doesn't blow up on them.

With that change, the n <= bytealg.MaxLen and n > bytealg.MaxLen cases are
now more similar, and this combines them. It also rearranges the fallbacks
so that we move on to the next search strategy by breaking out of a loop,
instead of nesting it in an if in the IndexByte loop. Though there's still
more code on net after the other changes, it may read more linearly now.

This is meant a starting point for trying to figure out whether this is a
good approach, whether it should be broken up, whether it causes important
regressions on non-amd64 platforms, what if any tests/benchmarks to add,
etc.